+++ components/Contact/Form/Form.module.css
@@ -1,0 +1,18 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+}
+
+.firstName,
+.lastName {
+  width: 100%;
+  padding-top: 20px;
+  margin-top: 16px;
+}
+
+/* input {
+  width: 75%;
+} */
