+++ package.json
@@ -12,6 +12,7 @@
     "@icons-pack/react-simple-icons": "^4.6.1",
     "next": "11.1.0",
     "react": "17.0.2",
+    "react-bootstrap": "^2.0.0-beta.5",
     "react-dom": "17.0.2",
     "react-icons": "^4.2.0",
     "react-modal": "^3.14.3",
